+++ delta/plugins/elasticsearch/src/main/scala/ch/epfl/bluebrain/nexus/delta/plugins/elasticsearch/routes/ElasticSearchViewsRoutes.scala
@@ -272,7 +272,7 @@ final class ElasticSearchViewsRoutes(
                             emit(
                               Created,
                               views
-                                .tag(id, ref, tag, tagRev.toInt, rev)
+                                .tag(id, ref, tag, tagRev, rev)
                                 .tapEval(index(ref, _, mode))
                                 .mapValue(_.metadata)
                                 .rejectWhen(decodingFailedOrViewNotFound)
